backend/pyproject.toml
["build-system"]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-demo-service"
version = "0.1.0"
description = "Standalone demo service for the AI requirement understanding & test-case generation engine."
requires-python = ">=3.10"
authors = [
    {name = "AI Module", email = "ai-demo@example.com"},
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "python-docx>=1.1.0",
    "python-multipart>=0.0.6",
    "sentence-transformers>=2.2.0",
    "chromadb>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.1.0",
    "ruff>=0.6.9",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "core*"]

[tool.ruff]
line-length = 110
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E203", "E501"]

